Added source testing (integration, interpolation).
The source now have some tests (for quads). They are place on a simple mesh, and it is required that their coordinates in real and reference space are correct. As well, a point source is fired once for each polynomial order from 1 -> 10, and it is asserted that they resulting interpolated field integrates to the correct value.