Added static ReceiverHDF5 functionality.
Now, when the HDF5 constructor is called, some static functions take care of the initialization of the HDF5 file. Simple and neat (thread safe?). As well, I've wrapped up the Receiver classes in their own std::unique_ptr. The motivation to do this came from testing (in each testing loop, we need to have all member variables destroyed), but also it should be good practice.