Updated clone function for receiver to use shared_ptr.
Having an internal debate on wheter to go through the trouble to use unique_ptrs where possible. This would required a bit of a re-factor of some of the class creations and destructions. It may be worth it though, and might be too memory safe to ignore. Anyways, things are std::shared_ptrs for now, and we'll see what we do about this in the future.