Fixed segfault when running older tests with new multi-field setup.
Now _all_ attachments of DOFs need some sort of model associated with them. This is because the DOFs themselves depend on mesh quantities. For example, if the element doesn't know whether it is acoustic or elastic, then how would it know how many DOFs to attach? I think this is reasonable. For example, even things like sources need to know the type of element (otherwise which direction would a certain force fire in?).