Added coverage to ExodusModel.
Took the opportunity to clean up some historical annoyances as well.
This being said, there's still some sketchy-ness. This is based mainly
around the current lack of standard on how we name our parameters. I
believe this can be easily fixed in the future, as soon as we write the
little mixin classes for isotropic / anisotropic propagation. Right now,
some value will be checked, and swapped if appropriate. If this happens,
the user will get a friendly warning.