Choose boundary types at runtime.
Added a new option to specify boundaries at runtime. Currently,
homogeneous Dirichlet and free surfaces are supported. To use, the user
would pass --homogeneous-dirichlet x0,x1,y0 to Salvus, which would make
those three interfaces Dirichlet boundaries. If a boundary is not
specified, it is by default a free surface. Notice that the
eigenfunction tests have been modified as well: we now mock the passing
of --homogeneous-dirichlet. Indeed, those tests fail if this is not
passed.